requirements.txt
numpy==1.21.5
pandas==1.5.3  # Atualizado
scikit-learn==1.0.2
flask==3.0.0
mlflow==1.23.1
joblib==0.16.0
lime==0.2.0.1
fluent-logger==0.10.0
bokeh==2.4.2
arrow==1.2.1
wickedhot==0.2.5
boto3==1.23.7
protobuf==3.20.1
psycopg2-binary==2.9.7
sqlalchemy==1.4.46  # Atualizado
# Testing Dependencies
requests-mock==1.8.0
flake8==3.7.8
flake8-print==3.1.4
autopep8==1.5
snakeviz==2.0.1
pytest==6.2.5
pytest-cov==3.0.0
coverage==6.2